refactor: remove old run_code_eval
